+++ source/extensions/omni.isaac.menu/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.0] - 2024-01-30
+### Changed
+- Changed sensor menu tests to verify sensor functionality through the sensor interfaces
+- Added golden image test for the environment
+
 ## [0.1.0] - 2024-01-19
 ### Added
 - Initial version of extension, moving menu code out of omni.isaac.utils extension